% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar2precision.R
\name{ar2precision}
\alias{ar2precision}
\title{Precision matrix for an AR2 model.}
\usage{
ar2precision(n, a)
}
\arguments{
\item{n}{the size of the model.}

\item{a}{the length three vector with the coefficients. See details.}
}
\value{
the precision matrix as a sparse matrix object with edge correction.
}
\description{
Creates a precision matrix as a sparse matrix object
considering the specification stated in the \section{details}.
}
\section{Details}{

Let the second order auto-regression model be defined as

\deqn{a_0 x_t + a_1 x_{t-1} + a_2 x_{t-2} = w_t, w_t ~ N(0, 1).}

The n times n symmetric precision matrix Q
for x_1, x_2, ..., x_n
has the following non-zero elements:

 \eqn{Q_{1,1} = Q_{n,n} = a_0^2}

 \eqn{Q_{2,2} = Q_{n-1,n-1} = a_0^2 + a_1^2}

 \eqn{Q_{1,2} = Q_{2,1} = Q_{n-1,n} = Q_{n,n-1} = a_0 a_1}

 \eqn{Q_{t,t} = q_0 = a_0^2 + a_1^2 + a_2^2, t = 3, 4, ..., n-2}

 \eqn{Q_{t,t-1} = Q_{t-1,t} = q_1 = a_1(a_0 + a_2), t = 3, 4, ..., n-1}

 \eqn{Q_{t,t-2} = Q_{t-2,t} = q_2 = a_2 a_0, t = 3, 4, ..., n}
}

\examples{
 ar2precision(7, c(1, -1.5, 0.9))
}
\seealso{
\link{ar2cor}
}
