% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar2precision.R
\name{ar2q}
\alias{ar2q}
\title{Illustrative code to build the precision matrix for an AR2 model.}
\usage{
ar2q(n, a)
}
\arguments{
\item{n}{the size of the model.}

\item{a}{the length three vector with the coefficients. See details.}
}
\value{
the precision matrix as a sparse matrix object with edge correction.
}
\description{
Creates a precision matrix as a sparse matrix object.
}
\section{Details}{

 Le the second order auto-regression model defined as
  a_0 x_t + a_1 x_{t-1} + a_2 x_{t-2} = w_t
where w_t ~ N(0, 1).
}

\seealso{
\link{ar2cor}
}
