% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlabru.R
\name{bru_get_mapper.stModel_cgeneric}
\alias{bru_get_mapper.stModel_cgeneric}
\alias{bru_get_mapper}
\title{Mapper object for automatic inlabru interface}
\usage{
bru_get_mapper.stModel_cgeneric(model, ...)
}
\arguments{
\item{model}{The model object (of class \code{stModel_cgeneric}, from
\code{stModel.define})}

\item{\dots}{Unused.}
}
\value{
A \code{bru_mapper} object of class \code{bru_mapper_multi} with
sub-mappers \code{space} and \code{time} based on the model \code{smesh} and \code{tmesh}
objects.
}
\description{
Return an \code{inlabru} \code{bru_mapper} object that can be used for computing
model matrices for the space-time model components.  The \code{bru_get_mapper()}
function is called by the \code{inlabru} methods to automatically obtain the
needed mapper object (from \code{inlabru} \verb{2.7.0.9001}; before that, use
\code{mapper = bru_get_mapper(model)} explicitly).
}
\seealso{
\code{\link[inlabru:bru_get_mapper]{inlabru::bru_get_mapper()}}
}
