% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgeneric_ast.R
\name{cgeneric_ast}
\alias{cgeneric_ast}
\title{Build an \code{cgeneric} conditional autoregression spacetime model.}
\usage{
cgeneric_ast(G, nt, control.priors, ...)
}
\arguments{
\item{nt}{integer to define the temporal dimention.}

\item{control.priors}{a named list with parameter priors,
named as \code{prs}, \code{prt} and \code{psigma}, each one as a vector
with length two containing (U, a) to define the
corresponding PC-prior such that, respectively,
P(range.spatial<U)=a, P(rho>U)=a or P(sigma>U)=a.
If a=0 or a=NA, then U is taken to be the fixed
value of the parameter.}

\item{...}{arguments (debug,useINLAprecomp,shlib)
passed on to \code{\link[INLAtools:cgeneric-class]{INLAtools::cgeneric()}}.}

\item{R}{a graph representing the spatial neighborhood.}
}
\value{
a \code{cgeneric} object, see \code{\link[INLAtools:cgeneric-class]{INLAtools::cgeneric()}}.
}
\description{
Build data needed to implement the model for the
C interface, \code{cgeneric}, in the 'INLA' package,
that can be used as a linear predictor
model component with an 'f' term.
}
\details{
The precision matrix is defined as
\deqn{Q = \tau ((D(\rho) + H(\rho)) \otimes (\kappa^2 I + R)}
where D and H are temporal matrices, and R is the
structure matrix supplied.
\eqn{\tau} is the (local) precision parameter,
\eqn{\kappa^2} is a scaling parameter and
\eqn{\rho} is the temporal autoregressive parameter.
}
\seealso{
\code{\link[INLAtools:cgeneric_get]{INLAtools::prior.cgeneric()}}
}
