% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgeneric_nngp.R
\name{cgeneric_nngp}
\alias{cgeneric_nngp}
\title{Build a \code{cgeneric} for Nearest Neighbour Gaussian Process.}
\usage{
cgeneric_nngp(dists, nnl, control.priors, correl = "Matern", nu = 1, ...)
}
\arguments{
\item{dists}{a \code{dist} object.}

\item{nnl}{a list with the neighbors indexes.}

\item{control.priors}{a named list with parameter priors,
named as \code{prs}, \code{prt} and \code{psigma}, each one as a vector
with length two containing (U, a) to define the
corresponding PC-prior such that, respectively,
P(range.spatial<U)=a, P(rho>U)=a or P(sigma>U)=a.
If a=0 or a=NA, then U is taken to be the fixed
value of the parameter.}

\item{correl}{character to specify the correlation function.
Available options are "Matern" (default) and "powerExponential".}

\item{nu}{numeric to specify the smoothness parameter.}

\item{...}{arguments (debug,useINLAprecomp,shlib)
passed on to \code{\link[INLAtools:cgeneric-class]{INLAtools::cgeneric()}}.}
}
\value{
a \code{cgeneric} object, see \code{\link[INLAtools:cgeneric-class]{INLAtools::cgeneric()}}.
}
\description{
Build data needed to implement the model for the
C interface, \code{cgeneric}, in the 'INLA' package,
that can be used as a linear predictor
model component with an 'f' term.
}
\details{
The precision matrix is defined as
\deqn{Q = LL',L = (D-A)D^{-0.5}}.
where A is a lower triangle matrix and D a diagonal one..
}
\seealso{
\code{\link[INLAtools:cgeneric_get]{INLAtools::prior.cgeneric()}}
}
