% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_package_version_and_load}
\alias{check_package_version_and_load}
\title{check package version and load}
\usage{
check_package_version_and_load(pkg, minimum_version, quietly = FALSE)
}
\arguments{
\item{pkg}{character with the name of the package}

\item{minimum_version}{character with the minimum required version}

\item{quietly}{logical indicating if messages shall be printed}
}
\description{
check package version and load
}
\note{
this is the same code as in the inlabru package
}
