% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcndUtils.R
\name{downloadUtilFiles}
\alias{downloadUtilFiles}
\alias{donwloadUtilFiles}
\title{Download files from the NOAA's GHCN daily data}
\usage{
downloadUtilFiles(data.dir, year = 2022, force = FALSE)
}
\arguments{
\item{data.dir}{the folder to store the files.}

\item{year}{the year of the daily weather data.}

\item{force}{logical indicating if it is to force
the download. If FALSE each file will be downloaded
if it does not exists locally yet.}
}
\value{
a named character vector with the local file names:
daily.data, stations.all, elevation.
}
\description{
Download files from the NOAA's GHCN daily data
}
