% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldMap.R
\name{Earth_poly}
\alias{Earth_poly}
\title{Function to define the boundary Earch polygon
in longlat projection for a given resolution.}
\usage{
Earth_poly(resol = 300, crs = "+proj=moll +units=km")
}
\arguments{
\item{resol}{is the number of subdivisions
along the latitude coordinates and half the
number of subdivisions along the longitude coordinates.}

\item{crs}{a string with the projection.
Default is the Mollweide projection with units in kilometers.}
}
\value{
a 'st_sfc' object with the Earth polygon.
}
\description{
Function to define the boundary Earch polygon
in longlat projection for a given resolution.
}
