% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcnDownload.R
\name{ghcnDownload}
\alias{ghcnDownload}
\title{Download daily data}
\usage{
ghcnDownload(year, local = ".")
}
\arguments{
\item{year}{an integer to specify the year}

\item{local}{the directory to save the downloaded file.
Default is the current working directory.}
}
\value{
it writes the downloaded file locally
}
\description{
This function helps to download data from NCDC/NOAA.
It first checks if the file exists locally.
}
\section{Warning}{

it takes time to download in case of bad internet connection
}

\seealso{
\code{\link[=gzVariableSelect]{gzVariableSelect()}}
}
