% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcndUtils.R
\name{ghcndSelect}
\alias{ghcndSelect}
\title{Select data from the daily dataset}
\usage{
ghcndSelect(
  gzfile,
  variable = c("TMIN", "TAVG", "TMAX"),
  qflag = "",
  verbose = TRUE,
  astype = as.integer
)
}
\arguments{
\item{gzfile}{the local filename for
the daily data file file. E.g. 2023.csv.gz from
\url{https://www.ncei.noaa.gov/pub/data/ghcn/daily/by_year/}
see references bellow.}

\item{variable}{string with the variable name(s) to be selected}

\item{qflag}{a string with quality control flag(s)}

\item{verbose}{logical indicating if progress is to be printed}
}
\value{
array \link{days, stations, variables} if more than one variable
or a matrix \link{days, stations} if one variable.
}
\description{
Select data from the daily dataset
}
\section{Details}{

The default selects TMIN, TAVG and TMAX and
return it as integer because the original data is also integer
with units in 10 Celcius degrees.
}

\section{Warning}{

It can take time to execute if, for example,
the data.table package is not available.
}

\references{
Menne, M., Durre, I., Vose, R., Gleason, B. and Houston, T. (2012)
An overview of the global historical climatology network-daily database.
Journal of Atmospheric and Oceanic Technology, 897–910.
}
