% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzVariableSelect.R
\name{gzVariableSelect}
\alias{gzVariableSelect}
\title{Select the specified variable}
\usage{
gzVariableSelect(
  gzfile,
  variable,
  astype = as.integer,
  qflag = "",
  verbose = TRUE
)
}
\arguments{
\item{gzfile}{the daily data file file. E.g. 2021.csv.gz from
\url{https://www.ncei.noaa.gov/pub/data/ghcn/daily/by_year/}
see references bellow.}

\item{variable}{a character indicating which variable to select.}

\item{astype}{function to be used to type the data.
Default is the current working directory.}

\item{qflag}{Extract data matching this quality flag. Default ''.}

\item{verbose}{logical to indicate if verbose mode is active}
}
\value{
array \link{days, stations, variables} if more than one variable
or a matrix \link{days, stations} if one variable.
}
\description{
Select a specivied variable from a .gz file and
save it into a .RData.
}
\section{Warning}{

it can take time to execute if the data.table package is not available.
}

\references{
Menne, M., Durre, I., Vose, R., Gleason, B. and Houston, T. (2012)
An overview of the global historical climatology network-daily database.
Journal of Atmospheric and Oceanic Technology, 897–910.
}
