% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Heron}
\alias{Heron}
\alias{Heron,}
\alias{Area,}
\alias{Stiffness}
\title{Internal util functions}
\usage{
Heron(x, y)
}
\arguments{
\item{x, y}{coordinate vectors,}
}
\value{
the area of a triangle
}
\description{
Internal util functions
}
\section{Warning}{
 Internal functions, not exported.
}

