% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLAspacetime.R
\name{INLAspacetime}
\alias{INLAspacetime}
\title{Functions for easier code to fit of some models with INLA.}
\usage{
INLAspacetime()
}
\value{
opens the Vignettes directory on a browser
}
\description{
This package main pourpose is to provide user friendly functions
to fit temporal, spatial and spatio-temporal models using the
INLA software available at www.r-inla.org
}
