% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stModel.matrices.R
\name{Jmatrices}
\alias{Jmatrices}
\title{The 2nd order temporal matrices with boundary correction}
\usage{
Jmatrices(tmesh)
}
\arguments{
\item{tmesh}{Temporal mesh}
}
\value{
return a list of temporal finite element method matrices
for the supplied mesh.
}
\description{
The 2nd order temporal matrices with boundary correction
}
\details{
Temporal GMRF representation with stationary boundary conditions
as in Appendix E of the paper.
}
