% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh2d.R
\name{mesh2d}
\alias{mesh2d}
\title{Illustrative code for building a mesh in 2d domain.}
\usage{
mesh2d(loc, domain, max.edge, offset, SP = TRUE)
}
\arguments{
\item{loc}{a two column matrix with location coordinates.}

\item{domain}{a two column matrix defining the domain.}

\item{max.edge}{the maximum edge length.}

\item{offset}{the length of the outer extension.}

\item{SP}{logical indicating if the output will include the SpatialPolygons.}
}
\value{
a mesh object.
}
\description{
Creates a mesh object. This is just a test code.
For efficient, reliable and general code
use the \pkg{fmesher} package.
}
\section{Warning}{

This is just for illustration purposes and one should consider the
efficient function \code{\link[fmesher:fm_mesh_2d]{fmesher::fm_mesh_2d()}} (and other related functions)
available a the fmesher package.
}

