% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceans50m.R
\name{oceans50m}
\alias{oceans50m}
\title{To retrieve the ocean map.}
\usage{
oceans50m(
  local = ".",
  zipfile = "oceans50m.zip",
  url = "https://stacks.stanford.edu/file/druid:jb593mf9286/data.zip"
)
}
\arguments{
\item{local}{the directory to save the downloaded file.
Default is the current working directory.}
}
\value{
the oceans map
}
\description{
Download, if not locally available, the ocean shapefile and
load it into the R environment.
}
\section{Warning}{

 it would not work if the file is not available
 locally and there is no internet connection.
}

\references{
\link{https://maps.princeton.edu/catalog/stanford-jb593mf9286}
}
\seealso{
\code{\link{worldMap}}
}
