% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde2precision.R
\name{spde2precision}
\alias{spde2precision}
\title{Illustrative code to build the precision matrix for SPDE kind models.}
\usage{
spde2precision(kappa, fem, alpha)
}
\arguments{
\item{kappa}{the scale parameter.}

\item{fem}{a list containing the Finite Element matrices.}

\item{alpha}{the smoothness parameter.}
}
\value{
the precision matrix as a sparse matrix object.
}
\description{
Creates a precision matrix as a sparse matrix object.
}
\section{Warning}{

 This is just for illustration purpose and one should consider the
 efficient function available a the INLA package.
}

