% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st121cgeneric.R
\name{st121cgeneric}
\alias{st121cgeneric}
\title{Define the spacetime model with alpha_t=1, alpha_s=2 and alpha_E=1.}
\usage{
st121cgeneric(tmesh, smesh, prior.rs, prior.rt, prior.sigma)
}
\arguments{
\item{tmesh}{the temporal mesh}

\item{smesh}{the spatial mesh}

\item{prior.rs}{a length two vector for the spatial practical range
prior paramters, as U and a so that P(rs<U)=a. If a=0,
then U is taken as, fixed, spatial range.}

\item{prior.rt}{a length two vector for the temporal practical
range prior paramters as U and a so that P(rt<U)=a.
If a=0 then U is take as, fixed, temporal range.}

\item{prior.sigma}{a length two vector for the sigma prior parameters,
as U and a so that P(sigma>U)=a. If a=0, then U is taken as, fixed, sigma.}
}
\value{
objects to be used in the cgeneric implementation
}
\description{
This function define the object set for using the 'cgeneric'
model tool in INLA.
}
\details{
See the paper.
}
