% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st121rgeneric.R
\name{st121rgeneric}
\alias{st121rgeneric}
\title{Define the spacetime model with alpha_t=1, alpha_s=2 and alpha_E=1.}
\usage{
st121rgeneric(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL,
  args = NULL,
  ...
)
}
\arguments{
\item{cmd}{see the 'rgeneric.define' in INLA.}

\item{theta}{hyperparameters of the model.}

\item{args}{arguments (for example, n: size of the model).}

\item{...}{objects for internal use.}
}
\value{
an rgeneric INLA model.
}
\description{
This function define the object set for using the 'rgeneric'
model tool in INLA.
}
\details{
See the paper.
}
