% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.inla.R
\name{stats.inla}
\alias{stats.inla}
\title{To retrieve goodness of fit statistics from an inla output object.}
\usage{
stats.inla(m, i = NULL, y, fsummarize = mean)
}
\arguments{
\item{m}{an inla output object.}

\item{i}{an index to subset the estimated values.}

\item{y}{observed to compare against.}

\item{fsummarize}{the summary function, default \code{mean}}
}
\value{
a vector with the extracted statistics
}
\description{
Exctracts dic, waic, po, cpo, mse, mae, crps and scrps.
}
\section{Warning}{

 The po, crps and scrps are computed using the posterior mean and the
 sum of the sd of the posterior fitted value plus the inverse of the
 likelihood precision posterior mean.
}

