% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUtils.R, R/utils.R
\name{stdSubs}
\alias{stdSubs}
\title{To check unusual low/high variance segments}
\usage{
stdSubs(x, nsub = 12, fs = 15)

stdSubs(x, nsub = 12, fs = 15)
}
\arguments{
\item{x}{numeric vector}

\item{nsub}{number for the segments length}

\item{fs}{numeric to use for detecting too
hight or too low local standard deviations.}
}
\value{
logical indicating if any of the
\code{st} are \code{fs} times lower/higher the average
of \code{st}, where is returned as an attribute:
\itemize{
\item attr(, 'st') numeric vector with the
standard deviation at each segment of the data.
}

logical indicating if any of the
\code{st} are \code{fs} times lower/higher the average
of \code{st}, where is returned as an attribute
ad detailed below.
}
\description{
To check unusual low/high variance segments

To check unusual low/high variance segments
}
\section{Attributes}{

\itemize{
\item attr(, 'st') numeric vector with the
standard deviation at each segment of the data.
}
}

