% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stModel.matrices.R
\name{stModel.precision}
\alias{stModel.precision}
\title{Spacetime precision matrix.}
\usage{
stModel.precision(smesh, tmesh, model, theta, verbose = FALSE)
}
\arguments{
\item{smesh}{a mesh object over the spatial domain.}

\item{tmesh}{a mesh object over the time domain.}

\item{model}{a string identifying the model.
So far we have the following models:
'102', '121', '202' and '220' models.}

\item{theta}{numeric vector of length three with
\eqn{log(gamma_s, gamma_t, gamma_e)}.}

\item{verbose}{logical to print intermediate objects.}
}
\value{
a (sparse) precision matrix,
as in Lindgren et. al. (2024)
}
\description{
To build the the precision matrix for a spacetime model
given the temporal and the spatial meshes.
}
