% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldMap.R
\name{world_grid}
\alias{world_grid}
\title{Define a regular grid in 'Mollweide' projection,
with units in kilometers.}
\usage{
world_grid(size = 50, domain)
}
\arguments{
\item{size}{the (in kilometers) of the grid cells.}

\item{domain}{if provided it should be an \code{sf} or \code{sfc} object.
In this case, the grid cells with centers falling inside
will be retrieved.}
}
\value{
a 'sf' points object with the centers of a
grid set within Earth (and the supplied domain)
}
\description{
Define a regular grid in 'Mollweide' projection,
with units in kilometers.
}
