% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldMap.R
\name{worldMap}
\alias{worldMap}
\title{Helper functions to retrieve the world map,
a world polygon, and create grid centers.}
\usage{
worldMap(
  crs = "+proj=moll +units=km",
  scale = "medium",
  returnclass = c("sf", "sv")
)
}
\arguments{
\item{crs}{a string with the projection.
Default is the Mollweide projection with units in kilometers.}

\item{scale}{see the help from the 'ne_countries'
of the 'rnaturalearth' package}

\item{returnclass}{defined by the
\link{sf} package if returnclass = "sf"
or \link{terra} if returnclass = "sv".}
}
\value{
the world map of the countries.
}
\description{
Retrieve the map of the countries
}
\references{
The land and ocean maps are obtained with
the 'rnaturalearth' package.
}
