% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzVariableSelect.R
\name{gzVariableSelect}
\alias{gzVariableSelect}
\title{Select the specified variable}
\usage{
gzVariableSelect(
  gzfile,
  variable,
  astype = as.integer,
  qflag = "",
  verbose = TRUE
)
}
\arguments{
\item{gzfile}{the .gz file}

\item{variable}{a character indicating which variable to select.}

\item{astype}{function to be used to type the data.
Default is the current working directory.}

\item{verbose}{logical to indicate if verbose mode is active}
}
\value{
it writes a new file
}
\description{
Select a specivied variable from a .gz file and
save it into a .RData.
}
\section{Warning}{

 it can take time to execute if the data.table package is not available.
}

\seealso{
\code{\link{ghcnDownload}}
}
