% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLAspacetime.R
\name{INLAspacetime}
\alias{INLAspacetime}
\title{Spatial and Spatio-Temporal Models using INLA}
\usage{
INLAspacetime()
}
\value{
opens the Vignettes directory on a browser
}
\description{
This package main purpose is to provide user friendly functions
to fit temporal, spatial and space-time models using the
INLA software available at www.r-inla.org as well the
inlabru package available
}
