% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.dual.R
\name{mesh.dual}
\alias{mesh.dual}
\title{Extracts the dual of a mesh object.}
\usage{
mesh.dual(mesh, SP = TRUE)
}
\arguments{
\item{mesh}{a 2d mesh object.}

\item{SP}{logical indicating if the output should be a SpatialPolygons object.}
}
\value{
a list of polygons or a SpatialPolygons object.
}
\description{
Creates a polygons list with the dual of a mesh object.
}
