% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.dual.R
\name{mesh.dual}
\alias{mesh.dual}
\title{Extracts the dual of a mesh object.}
\usage{
mesh.dual(
  mesh,
  returnclass = c("list", "sf", "SP"),
  mc.cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{mesh}{a 2d mesh object.}

\item{returnclass}{if
'list' return a list of polygon coordinates,
if "sf" return a sfc_multipolygon object,
if "SP" return a SpatialPolygons object.}

\item{mc.cores}{number of threads to be used.}
}
\value{
one of the three in 'returnclass'
}
\description{
Extracts the dual of a mesh object.
}
