% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh2fem.R
\name{mesh2fem}
\alias{mesh2fem}
\title{Illustrative code for Finite Element matrices of a mesh in 2d domain.}
\usage{
mesh2fem(mesh, order = 2)
}
\arguments{
\item{mesh}{a 2d mesh object.}

\item{order}{the desired order.}
}
\value{
a list object containing the FE matrices.
}
\description{
Creates a list of matrices object.
}
\section{Warning}{

This is just for illustration purpose and one should consider the
efficient function available a the INLA package.
}

