% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyUtils.R
\name{Heron}
\alias{Heron}
\alias{Area}
\alias{Stiffness}
\title{Internal util functions}
\usage{
Heron(x, y)

Area(x, y)

Stiffness(x, y)
}
\arguments{
\item{x, y}{coordinate vectors.}
}
\value{
the area of a triangle

the area of a general polygon

the stiffness matrix for a triangle
}
\description{
This computes the area of a triangle given its three coordinates.
}
\details{
Function used internally to compute the area of a triangle.
}
\section{Warning}{
 Internal functions, not exported.
}

