% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.inla.R
\name{stats.inla}
\alias{stats.inla}
\title{To retrieve goodness of fit statistics from an inla output object.}
\usage{
stats.inla(m, i = NULL, y, fsummarize = mean)
}
\arguments{
\item{m}{an inla output object.}

\item{i}{an index to subset the estimated values.}

\item{y}{observed to compare against.}

\item{fsummarize}{the summary function,
the default is \code{\link[base:mean]{base::mean()}}.}
}
\value{
a vector with the extracted statistics
}
\description{
Exctracts DIC, WAIC and CPO from an inla output and
computes PO, MSE, MAE, CRPS and SCRPS for a given input.
A summary is applied considering the user imputed function,
which by default is the mean. See Details and Warning!
}
\section{Details}{

When computing the PO, CRPS, and SCRPS scores it assumes a
Gaussian predictive distribution with mean equal to the
posterior mean and variance equal to the sum of the
posterior fitted value variance and the inverse of the
observation precision posterior mean.
}

\section{Warning}{

For CPO, PO, CRPS and SCRPS what is returned is the
negative of the applied summary.
}

