% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stModel.matrices.R
\name{stModel.matrices}
\alias{stModel.matrices}
\title{Define the spacetime model matrices.}
\usage{
stModel.matrices(smesh, tmesh, model)
}
\arguments{
\item{smesh}{a mesh object over the spatial domain.}

\item{tmesh}{a mesh object over the time domain.}

\item{model}{a string identifying the model.
So far we have the following models:
'102', '121', '202' and '220' models.}
}
\value{
a list containing
1. a string with the model identification
2. a length three vector with the constants `c1`, `c2` and `c3`
3. the vector `d`
4. the matrix `T`
5. the model matrices `M_1`, ..., `M_m`
}
\description{
This function computes all the matrices needed to build
the precision matrix for a non-stationary and non-separable
spatio-temporal model, as in ...
}
\details{
See the paper and the vignettes.
}
