% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperPadding.R
\name{upperPadding}
\alias{upperPadding}
\title{Prepare a matrix (or a list of matrices)
for use in a cgeneric code.}
\usage{
upperPadding(M)
}
\arguments{
\item{matrices}{a matrix a list of matrices}
}
\value{
If a unique matrix is given, return the
upper triangle considering the 'T' representation
in the Matrix package.
 If a list of matrices is given, return the graph of the
corresponding sum of all the matrices and the corresponding
values of the matrices.
}
\description{
The supplied matrix (or matrices) are prepared to
follow the required graph and precision pattern.
}
\examples{
A <- sparseMatrix(
   i=c(1,1,2,2,3,3,3,4,5,5),
   j=c(2,5,1,3,2,4,5,3,1,3), x=1)
A
upperPadding(A)
B <- Diagonal(5, colSums(A))
upperPadding(list(a=A, b=B))
}
