% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldMap.R
\name{worldMap}
\alias{worldMap}
\title{Helper functions to retrieve the world map,
a world polygon, and create grid centers.}
\usage{
worldMap(
  crs = "+proj=moll +units=km",
  scale = "medium",
  returnclass = c("sf", "sv")
)
}
\arguments{
\item{crs}{a string with the projection.
Default is the Mollweide projection
with units in kilometers.}

\item{scale}{The scale of map to return.
Please see the help of 'ne_countries' function
from the 'rnaturalearth' package.}

\item{returnclass}{A string determining the class
of the spatial object to return.
Please see the help of 'ne_countries' function
from the 'rnaturalearth' package.}
}
\description{
Retrieve the map of the countries
}
\references{
The land and ocean maps are obtained with
the 'rnaturalearth' package.
}
