% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldMap.R
\name{worldMap}
\alias{worldMap}
\title{Get the world map}
\usage{
worldMap(crs = sp::CRS("+proj=moll +units=km"))
}
\arguments{
\item{crs}{a CRS object for the target coordinates.
Defalt is the Mollweide projection with units in kilometers.}
}
\value{
the world map
}
\description{
Consider the world map, fix the wrap around and convert to the
Mollweide projection in km distance (default, can be changed).
}
\examples{
wmap <- worldMap()
sp::plot(wmap)
}
\references{
the world map data is from the maps R package.
}
